set hive type cost property from int to float; modify dockerfile

+++ src/graphql-schema/building/schema.js
@@ -15,7 +15,7 @@ const schema = `
 		id: ID!,
 		name: String,
 		demand: Float,
-		costs: Int
+		costs: Float
 	}
 
 	type Shop {
